requirements.txt
progress>=1.4
pyserial>=3.4
six>=1.12.0
xmodem>=0.4.5
Sphinx>=1.8.3
sphinx-argparse>=0.2.5
sphinxcontrib-documentedlist
appdirs
pygithub
bitstring
cython
kivy